Change position of increment statement
Initially, one could not insert a new node at the
end of the list since the loop would exit with i at a higher
value than the index of the last node. The condition allowing
a node to be inserted at the end would fail and hence no node
could be inserted. By changing the position of the increment
statement, the loop exits with the value of i corresponding
to the index of the last node.